agent/requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
openai
deepgram-sdk
pyngrok
python-dotenv
requests
